refactor(logReturns): extract into standalone domain module
