+++ domain/src/forecast/index.ts
@@ -1,3 +1,2 @@
 export * from "./forecast.derivers";
 export * from "./forecast.entities";
-export * from "./logReturnStats";
